Make new rectangles draw over old, regardless of color
This required restructuring the Firebase schema so that both types of
rectangles (and both types of pixels) go in the same database field.  I
also changed the database to avoid having to serialize coordinates into
strings, which simplified some of the handling code.

We now use push when adding drawings, which is cleaner.